Break up snap on ik
