+++ blender_source/MH_Community/operators/snaponikrig.py
@@ -2,7 +2,7 @@
 # -*- coding: utf-8 -*-
 
 import bpy
-from ..rig import RigInfo
+from ..rig import RigInfo, IkRig
 
 class SnapOnIkRigOperator(bpy.types.Operator):
     """Add bones which convert this to an IK Rig\n\nOnly Game or Kinect2 rigs."""
@@ -11,7 +11,6 @@ class SnapOnIkRigOperator(bpy.types.Operator):
     bl_options = {'REGISTER', 'UNDO'}
 
     def execute(self, context):
-        from ..snap_on_ik_rig import IkRig
         armature = context.object
 
         rigInfo = RigInfo.determineRig(armature)
