+++ blender_source/MH_Community/rig/fingerrig.py
@@ -1,4 +1,3 @@
-from .rig_info import *
 
 import bpy
 #===============================================================================
